pyproject.toml
[project]
name = "pretty-cfn"
version = "0.1.0"
description = "A CloudFormation YAML template formatter that aligns values for better readability"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "Your Name", email = "alessandro.bologna@gmail.com" }
]
dependencies = [
    "ruamel.yaml>=0.18.0",
    "click>=8.3.1",
    "pygments>=2.16.0",
    "cfn-lint>=1.41.0",
    "boto3>=1.41.0",
    "fastmcp>=2.13.1",
]

[project.scripts]
pretty-cfn = "pretty_cfn.main:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["pretty_cfn*"]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.0",
    "pytest-cov>=4.0",
    "ruff>=0.14.6"
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
